rename: 모델 분리
