+++ carea/lib/app/data/models/comment_model.dart
@@ -1,0 +1,30 @@
+import 'package:intl/intl.dart';
+
+class Comment {
+  int id;
+  int postId;
+  String content;
+  String created_at;
+  String nickname;
+
+  Comment({
+    required this.id,
+    required this.postId,
+    required this.content,
+    required this.created_at,
+    required this.nickname,
+  });
+
+  factory Comment.fromJson(Map<String, dynamic> json) {
+    DateTime createdAt = DateTime.parse(json['created_at']);
+    String formattedCreatedAt = DateFormat('yyyy-MM-dd').format(createdAt);
+
+    return Comment(
+      id: json['id'],
+      postId: json['post_id'],
+      content: json['content'],
+      created_at: formattedCreatedAt,
+      nickname: json['user']['nickname'],
+    );
+  }
+}
